refactored shared method
